+++ docs/src/components/Syntax.jsx
@@ -7,7 +7,31 @@ export default class Syntax extends React.Component {
   constructor(props) {
     super(props);
 
-    this.code = `
+    this.controlled = `
+require('codemirror/lib/codemirror.css');
+require('codemirror/theme/material.css');
+require('codemirror/theme/neat.css');
+require('codemirror/mode/xml/xml.js');
+require('codemirror/mode/javascript/javascript.js');
+
+import 'Codemirror' from 'react-codemirror2';
+
+<CodeMirror
+  value={this.state.value}
+  options={{
+    mode: this.mode
+    theme: this.theme,
+    lineNumbers: true
+  }}
+  onBeforeChange={(editor, data, value) => {
+    this.setState({value});
+  }}
+  onChange={(editor, value) => {
+    console.log('onChange#setState', {value});
+  }}
+/>`.trim();
+
+    this.uncontrolled = `
 require('codemirror/lib/codemirror.css');
 require('codemirror/theme/material.css');
 require('codemirror/theme/neat.css');
@@ -23,19 +47,15 @@ import 'Codemirror' from 'react-codemirror2';
     theme: this.theme,
     lineNumbers: true
   }}
-  onSet={(editor, value) => {
-    console.log('set', {value});
-  }}
-  onChange={(editor, metadata, value) => {
-    console.log('change', {value});
+  onChange={(editor, value) => {
+    console.log('onChange', {value});
   }}
 />`.trim();
-
   }
 
   tokenize() {
 
-    let code = prism.highlight(this.code, prism.languages.jsx);
+    let code = prism.highlight(this.props.controlled ? this.controlled : this.uncontrolled, prism.languages.jsx);
 
     return {
       __html: code