+++ test/index.spec.tsx
@@ -2,7 +2,6 @@ import * as React from 'react';
 import * as Adapter from 'enzyme-adapter-react-16';
 import * as Enzyme from 'enzyme';
 import * as sinon from 'sinon';
-
 import CodeMirror from '../src';
 
 Enzyme.configure({adapter: new Adapter()});
@@ -16,45 +15,6 @@ describe('CodeMirror: init', () => {
 
     expect.anything(mounted);
     expect.anything(mounted.html())
-  });
-});
-
-describe('CodeMirror: onSet', () => {
-
-  it('onSet(editor, value)', () => {
-    Enzyme.shallow(
-      <CodeMirror
-        value='foo'
-        onSet={(editor, value) => {
-          expect.anything(editor);
-          expect(value).toBe('foo');
-        }}/>
-    );
-  });
-});
-
-describe('CodeMirror: onBeforeSet', () => {
-
-  it('onBeforeSet(editor, value)', () => {
-
-    let callback;
-
-    let wrapper = Enzyme.shallow(
-      <CodeMirror
-        value='foo'
-        onBeforeSet={(editor, next) => {
-          callback = sinon.spy(next);
-          callback();
-        }}
-        onSet={(editor, value) => {
-          expect.anything(value);
-        }}
-        editorWillUnmount={(editor) => {
-          expect(callback.called).toBe(true);
-        }}/>
-    );
-
-    wrapper.unmount();
   });
 });
 
@@ -93,6 +53,74 @@ describe('CodeMirror: editorDidConfigure', () => {
         }}
         editorWillUnmount={(editor) => {
           expect(configured).toBe(true);
+          expect(callback.called).toBe(true);
+        }}/>
+    );
+
+    wrapper.unmount();
+  });
+});
+
+describe('CodeMirror: defineMode', () => {
+
+  it('defineMode', () => {
+
+    let mode = {
+      name: 'testMode',
+      fn: () => {
+        return {
+          startState: () => {
+          },
+          token: (stream, state) => {
+            stream.next();
+          }
+        }
+      }
+    };
+
+    let wrapper = Enzyme.shallow(
+      <CodeMirror
+        defineMode={mode}
+        editorDidMount={(editor, next) => {
+          expect(editor.doc.mode.name).toBe('testMode');
+        }}/>
+    )
+
+    wrapper.unmount();
+  });
+});
+
+describe('CodeMirror: onChange', () => {
+
+  it('onChange(editor, value)', () => {
+    Enzyme.shallow(
+      <CodeMirror
+        value='foo'
+        onChange={(editor, value) => {
+          expect.anything(editor);
+          expect(value).toBe('foo');
+        }}/>
+    );
+  });
+});
+
+describe('CodeMirror: onBeforeChange', () => {
+
+  it('onBeforeChange(editor, value)', () => {
+
+    let callback;
+
+    let wrapper = Enzyme.shallow(
+      <CodeMirror
+        value='foo'
+        onBeforeChange={(editor, next) => {
+          callback = sinon.spy(next);
+          callback();
+        }}
+        onChange={(editor, value) => {
+          expect.anything(value);
+        }}
+        editorWillUnmount={(editor) => {
           expect(callback.called).toBe(true);
         }}/>
     );